Training is not correct
I need to redo the training and make sure it is working how I want.
The stroke identifier is working, but it is registering the unknown
rune everytime, even when I draw a circle.  This probably has to do
with
how I train.  I don't think I am using the time data for trying,
but I think I should.