Transformer seems to be working better
I got rid of the noisy data and added another rune,
a cross.  So now between the cross and the circle it seems to be
able to tell them apart.

The next issue is the MAX_POINTS parameter.  That seems to be changing
a lot but it is necessary to put that in by hand later.  So, either I
figure out
how to automate extracting that value, or I fix it once and for all.